eventBridge

+++ lambda/invoices/invoiceImportFunction.ts
@@ -1,5 +1,5 @@
 import { APIGatewayProxyEvent, APIGatewayProxyResult, Context, S3Event, S3EventRecord } from "aws-lambda"
-import { ApiGatewayManagementApi, DynamoDB, S3 } from "aws-sdk"
+import { ApiGatewayManagementApi, DynamoDB, EventBridge, S3 } from "aws-sdk"
 import * as AWSXray from "aws-xray-sdk"
 import {v4 as uuid} from "uuid"
 import { InvoiceFile, InvoiceRepository } from "/opt/nodejs/invoiceRepository"
@@ -10,12 +10,14 @@ AWSXray.captureAWS(require("aws-sdk"))
 
 const invoicesDdb = process.env.INVOICE_DDB!
 const invoicesWsApiEndPoint = process.env.INVOICE_WSAPI_ENDPOINT!.substring(6)
+const AUDIT_BUS_NAME = process.env.AUDIT_BUS_NAME!
 
 const s3Client = new S3()
 const dbClient = new DynamoDB.DocumentClient()
 const apiGwManagementApi = new ApiGatewayManagementApi({
     endpoint: invoicesWsApiEndPoint
 })
+const eventBridgeClient = new EventBridge()
 
 const invoiceTransactionRepository = new InvoiceTransactionRepository(dbClient,invoicesDdb)
 const invoiceWSService = new InvoiceWSService(apiGwManagementApi)
@@ -59,31 +61,48 @@ async function processRecord(record: S3EventRecord): Promise<void> {
         const invoice = JSON.parse(objet.Body!.toString('utf-8'))  as InvoiceFile
         console.log(invoice)
 
-        const createInvoicePromisse = invoiceRepository.create({
-            pk: `#invoice_${invoice.customerName}`,
-            sk: invoice.invoiceNumber,
-            ttl: 0,
-            totalValue: invoice.totalValue,
-            productId: invoice.productId,
-            quantity: invoice.quantity,
-            transactionId: key,
-            createdAt: Date.now()
-        })
-
-        const deleteObjectPromise = s3Client.deleteObject({
-            Key: key,
-            Bucket: record.s3.bucket.name
-        }).promise()
-
-        const updateInvoicePromisse = await invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.PROCESSD)
-
-        const sendStatusPromisse = invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.PROCESSD)
-
-        await Promise.all([createInvoicePromisse, deleteObjectPromise, updateInvoicePromisse, sendStatusPromisse])
-
+        if(invoice.invoiceNumber.length >=5){
+            const createInvoicePromisse = invoiceRepository.create({
+                pk: `#invoice_${invoice.customerName}`,
+                sk: invoice.invoiceNumber,
+                ttl: 0,
+                totalValue: invoice.totalValue,
+                productId: invoice.productId,
+                quantity: invoice.quantity,
+                transactionId: key,
+                createdAt: Date.now()
+            })
+            const deleteObjectPromise = s3Client.deleteObject({
+                Key: key,
+                Bucket: record.s3.bucket.name
+            }).promise()
+            const updateInvoicePromisse = await invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.PROCESSD)
+            const sendStatusPromisse = invoiceWSService.sendInvoiceStatus(key, invoiceTransaction.connectionId, InvoiceTransactionStatus.PROCESSD)
+            await Promise.all([createInvoicePromisse, deleteObjectPromise, updateInvoicePromisse, sendStatusPromisse])
+        }else {
+            console.error(`invoice import failed - non valid invoice number -TransationId: ${key}`)
+            const putEventPromise = eventBridgeClient.putEvents(
+                {
+                    Entries: [
+                        {
+                            Source: 'app.invoice',
+                            EventBusName: AUDIT_BUS_NAME,
+                            DetailType: 'invoice',
+                            Time: new Date(),
+                            Detail: JSON.stringify({
+                                reason: 'FAIL_NO_INVOICE_NUMBER'
+                            })
+                        }
+                    ]
+                }    
+            ).promise()
+            const sendStatusPromise = invoiceWSService.sendInvoiceStatus(key,invoiceTransaction.connectionId, InvoiceTransactionStatus.NON_VALID_INVOICE_NUMBER)
+            const updateInvoicePromise = invoiceTransactionRepository.updateInvoiceTransaction(key, InvoiceTransactionStatus.NON_VALID_INVOICE_NUMBER)
+            await Promise.all([sendStatusPromise, updateInvoicePromise, putEventPromise])
+        }
+        await invoiceWSService.disconnectClient(invoiceTransaction.connectionId)
     } catch (error) {
         console.log((<Error>error).message)
     }
 
 }
-
